+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased ##
+
+### Added ###
+
+* Only one client is now allowed to be connected to the server at the same time. Extra clients will be rejected
+
 ## [v0.1.0-indev01] - 2022-11-06 ##
 
 [v0.1.0-indev01]: https://github.com/mfederczuk/usockit/releases/tag/v0.1.0-indev01